First iteration works, but fit is wrong for 2nd
Seems to be an issue with the potrs function. That thing is
already kind of set up weird, so I will try to just replace it
with how we did it in ReFacTo. This shouldn't make a huge difference
since this operation accounts for a small portion of the CPD time.